Organize searchbox files
