+++ src/components/node/NodePreview.vue
