+++ src/components/node/NodeSourceChip.vue
