+++ src/components/searchbox/NodeSearchBoxPopover.vue
