+++ src/components/searchbox/NodeSearchFilter.vue
